backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Compress==1.14.0
Flask-Caching==2.1.0
openai>=1.0.0
httpx>=0.27.0
pytest>=8.3.0
# google-generativeai is optional; install separately if you use Gemini embeddings
# google-generativeai>=0.3.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.9
dateparser>=1.2.0
sendgrid>=6.11.0
Pillow>=10.0.0
sentence-transformers>=3.2.0
